i3/toggle_touchpad.sh: awk script for touchpad autodetection
Previously I used a constant number to refer to the touchpad in the
toggling script. But since it has to different numbers assigned in
`xinput` for my two laptops, I made the script more general.

Now a little awk scripts is looking up the line containing *touchPad*
and extract the corresponding input number. This one is then used
throughout the script.